[lightify] Move config classes to internal.
